refactor: modularize the Toolbox runtime
