+++ src/companion-kernel/lib.rs
@@ -1,206 +1,15 @@
 #![no_std]
 
-use core::mem::size_of;
 use core::panic::PanicInfo;
 use core::ptr::{read_volatile, write_volatile};
 
-const SNAPSHOT_BYTES: u32 = size_of::<Snapshot>() as u32;
-const CONFIG_BYTES: u32 = size_of::<Layout>() as u32;
-const MAGIC: u32 = 0x4254_5747;
-const ABI_AND_SIZE: u32 = (SNAPSHOT_BYTES << 16) | 1;
+mod abi;
+mod cursor;
+mod memory;
+mod toolbox;
 
-const FLAG_READY: u32 = 1 << 0;
-const FLAG_PLAYER_VALID: u32 = 1 << 1;
-const FLAG_TARGET_VALID: u32 = 1 << 2;
-const FLAG_LOADING: u32 = 1 << 3;
-
-const FEATURE_NATIVE_CURSOR: u32 = 1 << 0;
-const FEATURE_TARGET_READOUT: u32 = 1 << 1;
-const FEATURE_TOOLBOX_FOUNDATION: u32 = 1 << 2;
-const KNOWN_FEATURES: u32 =
-    FEATURE_NATIVE_CURSOR | FEATURE_TARGET_READOUT | FEATURE_TOOLBOX_FOUNDATION;
-
-const TOOLBOX_BYTES: u32 = size_of::<ToolboxSnapshot>() as u32;
-const TOOLBOX_MAGIC: u32 = 0x5854_5747;
-const TOOLBOX_ABI_AND_SIZE: u32 = (TOOLBOX_BYTES << 16) | 1;
-const FLAG_HERO_AVAILABLE: u32 = 1 << 0;
-
-const DISPATCH_TICK: u32 = 0;
-const DISPATCH_CURSOR: u32 = 1;
-const DISPATCH_UI: u32 = 2;
-const PANEL_UNKNOWN: u32 = 0;
-const PANEL_HIDDEN: u32 = 1;
-const PANEL_SHOWN: u32 = 2;
-const COMMAND_IDLE: u32 = 0;
-const COMMAND_APPLIED: u32 = 1;
-const COMMAND_UNAVAILABLE: u32 = 2;
-
-const CURSOR_BYTES: u32 = size_of::<CursorSnapshot>() as u32;
-const CURSOR_MAGIC: u32 = 0x4354_5747;
-const CURSOR_ABI_AND_SIZE: u32 = (CURSOR_BYTES << 16) | 1;
-
-const FLAG_CURSOR_VALID: u32 = 1 << 0;
-const FLAG_CURSOR_HIDDEN: u32 = 1 << 1;
-const FLAG_CURSOR_UNSUPPORTED: u32 = 1 << 2;
-
-const CURSOR_EDGE: u32 = 32;
-const CURSOR_WORDS: u32 = CURSOR_EDGE * CURSOR_EDGE;
-const CURSOR_PIXEL_BYTES: u32 = CURSOR_WORDS * 4;
-// 'grtx', the texture handle's access key.
-const CURSOR_TEXTURE_KEY: u32 = 0x6772_7478;
-const CURSOR_TEXTURE_TYPE: u32 = 10;
-
-#[repr(C)]
-#[derive(Clone, Copy)]
-struct Layout {
-    context_root: u32,
-    agent_array: u32,
-    manual_target_agent_id: u32,
-    automatic_target_agent_id: u32,
-    game_context_slot: u32,
-    character_context: u32,
-    map_id: u32,
-    is_explorable: u32,
-    current_map_id: u32,
-    current_instance_type: u32,
-    player_number: u32,
-    agent_id: u32,
-    agent_x: u32,
-    agent_y: u32,
-    agent_type: u32,
-    agent_player_number: u32,
-    agent_model_type: u32,
-    cursor_active_art: u32,
-    cursor_software_model: u32,
-    cursor_show_count: u32,
-    cursor_color_buffer: u32,
-    cursor_art_hotspot: u32,
-    cursor_art_texture: u32,
-    cursor_handle_key: u32,
-    cursor_handle_object: u32,
-    cursor_view_texture: u32,
-    cursor_texture_type: u32,
-    cursor_texture_width: u32,
-    cursor_texture_height: u32,
-    party_context: u32,
-    player_party: u32,
-    party_heroes: u32,
-    hero_member_stride: u32,
-    hero_agent_id: u32,
-    hero_owner_player_id: u32,
-    hero_id: u32,
-    prop_context_slot: u32,
-    player_chat_message: u32,
-    hide_hero_panel_message: u32,
-    show_hero_panel_message: u32,
-}
-
-impl Layout {
-    const EMPTY: Self = Self {
-        context_root: 0,
-        agent_array: 0,
-        manual_target_agent_id: 0,
-        automatic_target_agent_id: 0,
-        game_context_slot: 0,
-        character_context: 0,
-        map_id: 0,
-        is_explorable: 0,
-        current_map_id: 0,
-        current_instance_type: 0,
-        player_number: 0,
-        agent_id: 0,
-        agent_x: 0,
-        agent_y: 0,
-        agent_type: 0,
-        agent_player_number: 0,
-        agent_model_type: 0,
-        cursor_active_art: 0,
-        cursor_software_model: 0,
-        cursor_show_count: 0,
-        cursor_color_buffer: 0,
-        cursor_art_hotspot: 0,
-        cursor_art_texture: 0,
-        cursor_handle_key: 0,
-        cursor_handle_object: 0,
-        cursor_view_texture: 0,
-        cursor_texture_type: 0,
-        cursor_texture_width: 0,
-        cursor_texture_height: 0,
-        party_context: 0,
-        player_party: 0,
-        party_heroes: 0,
-        hero_member_stride: 0,
-        hero_agent_id: 0,
-        hero_owner_player_id: 0,
-        hero_id: 0,
-        prop_context_slot: 0,
-        player_chat_message: 0,
-        hide_hero_panel_message: 0,
-        show_hero_panel_message: 0,
-    };
-}
-
-#[repr(C)]
-#[derive(Clone, Copy)]
-struct Snapshot {
-    magic: u32,
-    abi_and_size: u32,
-    sequence: u32,
-    flags: u32,
-    tick_count: u32,
-    map_id: u32,
-    instance_type: u32,
-    player_id: u32,
-    player_x: f32,
-    player_y: f32,
-    target_id: u32,
-    target_type: u32,
-    target_x: f32,
-    target_y: f32,
-    distance: f32,
-    range_band: u32,
-}
-
-// Separate bounded region: the 64-byte core snapshot is full, and the cursor
-// bitmap is far too large to live in it.
-#[repr(C)]
-struct CursorSnapshot {
-    magic: u32,
-    abi_and_size: u32,
-    sequence: u32,
-    flags: u32,
-    generation: u32,
-    width: u32,
-    height: u32,
-    hotspot_x: u32,
-    hotspot_y: u32,
-    pixel_hash: u32,
-    reserved: [u32; 6],
-    pixels: [u32; 1024],
-}
-
-#[repr(C)]
-struct ToolboxSnapshot {
-    magic: u32,
-    abi_and_size: u32,
-    sequence: u32,
-    flags: u32,
-    player_chat_count: u32,
-    cursor_event_count: u32,
-    hero_count: u32,
-    first_hero_id: u32,
-    first_hero_agent_id: u32,
-    panel_state: u32,
-    command_request: u32,
-    command_complete: u32,
-    command_status: u32,
-    reserved: [u32; 3],
-}
-
-const _: [(); 160] = [(); size_of::<Layout>()];
-const _: [(); 64] = [(); size_of::<Snapshot>()];
-const _: [(); 4160] = [(); size_of::<CursorSnapshot>()];
-const _: [(); 64] = [(); size_of::<ToolboxSnapshot>()];
+use abi::*;
+use memory::*;
 
 static mut SNAPSHOT_PTR: u32 = 0;
 static mut LAYOUT: Layout = Layout::EMPTY;
@@ -208,23 +17,6 @@ static mut INITIALIZED: bool = false;
 static mut FEATURES: u32 = 0;
 static mut TICK_COUNT: u32 = 0;
 static mut SEQUENCE: u32 = 0;
-static mut CURSOR_PTR: u32 = 0;
-static mut CURSOR_SEQUENCE: u32 = 0;
-static mut CURSOR_GENERATION: u32 = 0;
-static mut CURSOR_PUBLISHED: CursorPublished = CursorPublished::EMPTY;
-static mut CURSOR_DIRTY: bool = true;
-static mut CURSOR_EVENT_COUNT: u32 = 0;
-static mut TOOLBOX_PTR: u32 = 0;
-static mut TOOLBOX_SEQUENCE: u32 = 0;
-static mut PLAYER_CHAT_COUNT: u32 = 0;
-static mut HERO_COUNT: u32 = 0;
-static mut FIRST_HERO_ID: u32 = 0;
-static mut FIRST_HERO_AGENT_ID: u32 = 0;
-static mut PANEL_STATE: u32 = PANEL_UNKNOWN;
-static mut PANEL_PENDING: u32 = PANEL_UNKNOWN;
-static mut COMMAND_REQUEST: u32 = 0;
-static mut COMMAND_COMPLETE: u32 = 0;
-static mut COMMAND_STATUS: u32 = COMMAND_IDLE;
 
 #[link(wasm_import_module = "game")]
 extern "C" {
@@ -239,61 +31,6 @@ extern "C" {
 #[panic_handler]
 fn panic(_info: &PanicInfo) -> ! {
     loop {}
-}
-
-fn memory_bytes() -> u32 {
-    core::arch::wasm32::memory_size(0)
-        .saturating_mul(65_536)
-        .min(u32::MAX as usize) as u32
-}
-
-fn checked_add(left: u32, right: u32) -> Option<u32> {
-    left.checked_add(right)
-}
-
-fn checked_mul(left: u32, right: u32) -> Option<u32> {
-    left.checked_mul(right)
-}
-
-fn offset(base: u32, field: u32) -> Option<u32> {
-    checked_add(base, field)
-}
-
-fn indexed(base: u32, index: u32, stride: u32) -> Option<u32> {
-    checked_add(base, checked_mul(index, stride)?)
-}
-
-fn contains(address: u32, bytes: u32) -> bool {
-    checked_add(address, bytes).is_some_and(|end| end <= memory_bytes())
-}
-
-fn valid_region(enabled: bool, address: u32, bytes: u32, expected: u32) -> bool {
-    if enabled {
-        address != 0 && bytes == expected && address & 3 == 0 && contains(address, bytes)
-    } else {
-        address == 0 && bytes == 0
-    }
-}
-
-unsafe fn read_u32(address: u32) -> Option<u32> {
-    contains(address, 4).then(|| unsafe { read_volatile(address as *const u32) })
-}
-
-unsafe fn read_i32(address: u32) -> Option<i32> {
-    contains(address, 4).then(|| unsafe { read_volatile(address as *const i32) })
-}
-
-unsafe fn read_u16(address: u32) -> Option<u16> {
-    contains(address, 2).then(|| unsafe { read_volatile(address as *const u16) })
-}
-
-unsafe fn read_f32(address: u32) -> Option<f32> {
-    contains(address, 4).then(|| unsafe { read_volatile(address as *const f32) })
-}
-
-unsafe fn pointer(address: u32, required_bytes: u32) -> Option<u32> {
-    let value = unsafe { read_u32(address)? };
-    (value & 3 == 0 && contains(value, required_bytes)).then_some(value)
 }
 
 fn finite_position(value: f32) -> bool {
@@ -374,12 +111,7 @@ impl State {
     }
 }
 
-unsafe fn read_agent(
-    layout: Layout,
-    agent_buffer: u32,
-    size: u32,
-    id: u32,
-) -> Option<AgentState> {
+unsafe fn read_agent(layout: Layout, agent_buffer: u32, size: u32, id: u32) -> Option<AgentState> {
     if id == 0 || id >= size {
         return None;
     }
@@ -418,9 +150,8 @@ unsafe fn collect(layout: Layout) -> State {
         None => return state,
     };
 
-    let read_character = |field| {
-        offset(character, field).and_then(|address| unsafe { read_u32(address) })
-    };
+    let read_character =
+        |field| offset(character, field).and_then(|address| unsafe { read_u32(address) });
     let Some(base_map) = read_character(layout.map_id) else {
         return state;
     };
@@ -472,11 +203,7 @@ unsafe fn collect(layout: Layout) -> State {
     let Some(agent_bytes) = checked_mul(size, 4) else {
         return state;
     };
-    if size == 0
-        || size > capacity
-        || capacity > 4_096
-        || !contains(agent_buffer, agent_bytes)
-    {
+    if size == 0 || size > capacity || capacity > 4_096 || !contains(agent_buffer, agent_bytes) {
         return state;
     }
 
@@ -490,24 +217,19 @@ unsafe fn collect(layout: Layout) -> State {
         let Some(id_address) = offset(agent_address, layout.agent_id) else {
             return state;
         };
-        let Some(player_number_address) =
-            offset(agent_address, layout.agent_player_number)
-        else {
+        let Some(player_number_address) = offset(agent_address, layout.agent_player_number) else {
             return state;
         };
-        let Some(model_type_address) = offset(agent_address, layout.agent_model_type)
-        else {
+        let Some(model_type_address) = offset(agent_address, layout.agent_model_type) else {
             return state;
         };
         if unsafe { read_u32(id_address) } != Some(id)
             || unsafe { read_u16(player_number_address) } != Some(player_number as u16)
-            || unsafe { read_u16(model_type_address) }.map(|value| value & 0xf000)
-                != Some(0x3000)
+            || unsafe { read_u16(model_type_address) }.map(|value| value & 0xf000) != Some(0x3000)
         {
             continue;
         }
-        let Some(player) = (unsafe { read_agent(layout, agent_buffer, size, id) })
-        else {
+        let Some(player) = (unsafe { read_agent(layout, agent_buffer, size, id) }) else {
             return state;
         };
         state.flags = FLAG_READY | FLAG_PLAYER_VALID;
@@ -522,16 +244,12 @@ unsafe fn collect(layout: Layout) -> State {
         return State::empty();
     }
 
-    if layout.manual_target_agent_id != 0
-        && layout.automatic_target_agent_id != 0
-    {
+    if layout.manual_target_agent_id != 0 && layout.automatic_target_agent_id != 0 {
         let target_id = unsafe { read_u32(layout.manual_target_agent_id) }
             .filter(|id| *id != 0)
             .or_else(|| unsafe { read_u32(layout.automatic_target_agent_id) });
         if let Some(target_id) = target_id {
-            if let Some(target) =
-                unsafe { read_agent(layout, agent_buffer, size, target_id) }
-            {
+            if let Some(target) = unsafe { read_agent(layout, agent_buffer, size, target_id) } {
                 let dx = target.x - state.player.x;
                 let dy = target.y - state.player.y;
                 let distance_squared = dx * dx + dy * dy;
@@ -548,8 +266,7 @@ unsafe fn collect(layout: Layout) -> State {
 }
 
 unsafe fn collect_first_owned_hero(layout: Layout, state: State) -> (u32, u32, u32) {
-    if state.flags & (FLAG_READY | FLAG_PLAYER_VALID)
-        != (FLAG_READY | FLAG_PLAYER_VALID)
+    if state.flags & (FLAG_READY | FLAG_PLAYER_VALID) != (FLAG_READY | FLAG_PLAYER_VALID)
         || state.game == 0
         || state.player_number == 0
         || layout.hero_member_stride < 12
@@ -613,18 +330,17 @@ unsafe fn collect_first_owned_hero(layout: Layout, state: State) -> (u32, u32, u
         let Some(member) = indexed(buffer, index, layout.hero_member_stride) else {
             return (0, 0, 0);
         };
-        let Some(owner) = offset(member, layout.hero_owner_player_id)
-            .and_then(|at| unsafe { read_u32(at) })
+        let Some(owner) =
+            offset(member, layout.hero_owner_player_id).and_then(|at| unsafe { read_u32(at) })
         else {
             return (0, 0, 0);
         };
-        let Some(hero_id) = offset(member, layout.hero_id)
-            .and_then(|at| unsafe { read_u32(at) })
+        let Some(hero_id) = offset(member, layout.hero_id).and_then(|at| unsafe { read_u32(at) })
         else {
             return (0, 0, 0);
         };
-        let Some(agent_id) = offset(member, layout.hero_agent_id)
-            .and_then(|at| unsafe { read_u32(at) })
+        let Some(agent_id) =
+            offset(member, layout.hero_agent_id).and_then(|at| unsafe { read_u32(at) })
         else {
             return (0, 0, 0);
         };
@@ -645,83 +361,6 @@ unsafe fn collect_first_owned_hero(layout: Layout, state: State) -> (u32, u32, u
         count += 1;
     }
     (count, first_id, first_agent)
-}
-
-unsafe fn publish_toolbox() {
-    let next = unsafe { TOOLBOX_SEQUENCE }.wrapping_add(2) & !1;
-    let snapshot = unsafe { TOOLBOX_PTR as *mut ToolboxSnapshot };
-    let flags = if unsafe { FIRST_HERO_ID } != 0 {
-        FLAG_HERO_AVAILABLE
-    } else {
-        0
-    };
-    unsafe {
-        write_volatile(&mut (*snapshot).sequence, next.wrapping_sub(1));
-        write_volatile(&mut (*snapshot).magic, TOOLBOX_MAGIC);
-        write_volatile(&mut (*snapshot).abi_and_size, TOOLBOX_ABI_AND_SIZE);
-        write_volatile(&mut (*snapshot).flags, flags);
-        write_volatile(&mut (*snapshot).player_chat_count, PLAYER_CHAT_COUNT);
-        write_volatile(&mut (*snapshot).cursor_event_count, CURSOR_EVENT_COUNT);
-        write_volatile(&mut (*snapshot).hero_count, HERO_COUNT);
-        write_volatile(&mut (*snapshot).first_hero_id, FIRST_HERO_ID);
-        write_volatile(&mut (*snapshot).first_hero_agent_id, FIRST_HERO_AGENT_ID);
-        write_volatile(&mut (*snapshot).panel_state, PANEL_STATE);
-        write_volatile(&mut (*snapshot).command_request, COMMAND_REQUEST);
-        write_volatile(&mut (*snapshot).command_complete, COMMAND_COMPLETE);
-        write_volatile(&mut (*snapshot).command_status, COMMAND_STATUS);
-        write_volatile(&mut (*snapshot).sequence, next);
-        TOOLBOX_SEQUENCE = next;
-    }
-}
-
-unsafe fn tick_foundation(layout: Layout, state: State) {
-    let (count, hero_id, agent_id) = unsafe { collect_first_owned_hero(layout, state) };
-    unsafe {
-        if FIRST_HERO_ID != hero_id {
-            PANEL_STATE = PANEL_UNKNOWN;
-        }
-        HERO_COUNT = count;
-        FIRST_HERO_ID = hero_id;
-        FIRST_HERO_AGENT_ID = agent_id;
-    }
-    unsafe { apply_pending_hero_command(layout, state) };
-    unsafe { publish_toolbox() };
-}
-
-// Internal calls made after the game tick need the same PropContext that the
-// official browser client installs around its own calls. Save the slot, install
-// the already-validated GameContext for this one synchronous dispatch, and
-// restore the exact prior value before publishing completion.
-unsafe fn apply_pending_hero_command(layout: Layout, state: State) {
-    let desired = unsafe { PANEL_PENDING };
-    if desired == PANEL_UNKNOWN {
-        return;
-    }
-    unsafe { PANEL_PENDING = PANEL_UNKNOWN };
-    let hero_id = unsafe { FIRST_HERO_ID };
-    if hero_id == 0 || state.game == 0 {
-        unsafe {
-            COMMAND_COMPLETE = COMMAND_REQUEST;
-            COMMAND_STATUS = COMMAND_UNAVAILABLE;
-        }
-        return;
-    }
-    let message = if desired == PANEL_SHOWN {
-        layout.show_hero_panel_message
-    } else {
-        layout.hide_hero_panel_message
-    };
-    let previous = unsafe { read_volatile(layout.prop_context_slot as *const u32) };
-    unsafe {
-        write_volatile(layout.prop_context_slot as *mut u32, state.game);
-        ui_original(message, hero_id, 0);
-        write_volatile(layout.prop_context_slot as *mut u32, previous);
-    }
-    unsafe {
-        PANEL_STATE = desired;
-        COMMAND_COMPLETE = COMMAND_REQUEST;
-        COMMAND_STATUS = COMMAND_APPLIED;
-    }
 }
 
 unsafe fn publish(state: State) {
@@ -746,201 +385,6 @@ unsafe fn publish(state: State) {
         write_volatile(&mut (*snapshot).range_band, state.band);
         write_volatile(&mut (*snapshot).sequence, next);
         SEQUENCE = next;
-    }
-}
-
-#[derive(Clone, Copy)]
-struct CursorState {
-    hash: u32,
-    hotspot_x: u32,
-    hotspot_y: u32,
-    hidden: bool,
-    source: u32,
-}
-
-// The published identity. The active art pointer is not stable across cursor
-// changes, so the pixel hash is the only usable change key.
-#[derive(Clone, Copy, PartialEq)]
-struct CursorPublished {
-    flags: u32,
-    hash: u32,
-    hotspot_x: u32,
-    hotspot_y: u32,
-}
-
-impl CursorPublished {
-    const EMPTY: Self = Self {
-        flags: 0,
-        hash: 0,
-        hotspot_x: 0,
-        hotspot_y: 0,
-    };
-}
-
-// FNV-1a over the source BGRA words, so an unchanged cursor costs one pass and
-// no conversion. None means unreadable or never committed by the game.
-unsafe fn hash_cursor_pixels(source: u32) -> Option<u32> {
-    let mut hash: u32 = 0x811c_9dc5;
-    let mut committed: u32 = 0;
-    for index in 0..CURSOR_WORDS {
-        let word = unsafe { read_u32(indexed(source, index, 4)?)? };
-        hash = (hash ^ word).wrapping_mul(0x0100_0193);
-        committed |= word;
-    }
-    (committed != 0).then_some(hash)
-}
-
-// The readback that fills the colour buffer uses a hard-coded pitch, so a
-// source texture that is not 32x32 would have misfilled it.
-unsafe fn read_cursor(layout: Layout) -> Option<CursorState> {
-    let art = unsafe { pointer(layout.cursor_active_art, 24)? };
-    let handle = unsafe { pointer(offset(art, layout.cursor_art_texture)?, 12)? };
-    if unsafe { read_u32(offset(handle, layout.cursor_handle_key)?)? }
-        != CURSOR_TEXTURE_KEY
-    {
-        return None;
-    }
-    let view = unsafe { pointer(offset(handle, layout.cursor_handle_object)?, 12)? };
-    let texture = unsafe { pointer(offset(view, layout.cursor_view_texture)?, 0x68)? };
-    if unsafe { read_u32(offset(texture, layout.cursor_texture_type)?)? }
-        != CURSOR_TEXTURE_TYPE
-        || unsafe { read_u32(offset(texture, layout.cursor_texture_width)?)? }
-            != CURSOR_EDGE
-        || unsafe { read_u32(offset(texture, layout.cursor_texture_height)?)? }
-            != CURSOR_EDGE
-    {
-        return None;
-    }
-
-    let hotspot = offset(art, layout.cursor_art_hotspot)?;
-    let hotspot_x = unsafe { read_u32(hotspot)? };
-    let hotspot_y = unsafe { read_u32(offset(hotspot, 4)?)? };
-    if hotspot_x >= CURSOR_EDGE || hotspot_y >= CURSOR_EDGE {
-        return None;
-    }
-
-    let source = layout.cursor_color_buffer;
-    if !contains(source, CURSOR_PIXEL_BYTES) {
-        return None;
-    }
-    let hash = unsafe { hash_cursor_pixels(source)? };
-    let hidden = unsafe { read_i32(layout.cursor_show_count) }
-        .is_some_and(|count| count < 0);
-    Some(CursorState {
-        hash,
-        hotspot_x,
-        hotspot_y,
-        hidden,
-        source,
-    })
-}
-
-unsafe fn collect_cursor(layout: Layout) -> Result<CursorState, u32> {
-    if unsafe { read_u32(layout.cursor_software_model) } != Some(0) {
-        return Err(FLAG_CURSOR_UNSUPPORTED);
-    }
-    unsafe { read_cursor(layout) }.ok_or(0)
-}
-
-// `source` is None for a header-only update: it clears CURSOR_VALID without
-// disturbing the last good pixels.
-unsafe fn publish_cursor(published: CursorPublished, source: Option<u32>) {
-    let next = unsafe { CURSOR_SEQUENCE }.wrapping_add(2) & !1;
-    let cursor = unsafe { CURSOR_PTR as *mut CursorSnapshot };
-    unsafe {
-        write_volatile(&mut (*cursor).sequence, next.wrapping_sub(1));
-        write_volatile(&mut (*cursor).magic, CURSOR_MAGIC);
-        write_volatile(&mut (*cursor).abi_and_size, CURSOR_ABI_AND_SIZE);
-        write_volatile(&mut (*cursor).flags, published.flags);
-    }
-    if let Some(source) = source {
-        unsafe {
-            CURSOR_GENERATION = CURSOR_GENERATION.wrapping_add(1);
-            write_volatile(&mut (*cursor).generation, CURSOR_GENERATION);
-            write_volatile(&mut (*cursor).width, CURSOR_EDGE);
-            write_volatile(&mut (*cursor).height, CURSOR_EDGE);
-            write_volatile(&mut (*cursor).hotspot_x, published.hotspot_x);
-            write_volatile(&mut (*cursor).hotspot_y, published.hotspot_y);
-            write_volatile(&mut (*cursor).pixel_hash, published.hash);
-        }
-        for index in 0..CURSOR_WORDS {
-            let word = indexed(source, index, 4)
-                .and_then(|address| unsafe { read_u32(address) })
-                .unwrap_or(0);
-            // BGRA -> RGBA: keep alpha and green, swap red and blue.
-            let rgba =
-                (word & 0xff00_ff00) | ((word >> 16) & 0xff) | ((word & 0xff) << 16);
-            unsafe { write_volatile(&mut (*cursor).pixels[index as usize], rgba) };
-        }
-    }
-    unsafe {
-        write_volatile(&mut (*cursor).sequence, next);
-        CURSOR_SEQUENCE = next;
-        CURSOR_PUBLISHED = published;
-    }
-}
-
-unsafe fn tick_cursor(layout: Layout) {
-    if !unsafe { CURSOR_DIRTY } {
-        let last = unsafe { CURSOR_PUBLISHED };
-        if last.flags & FLAG_CURSOR_VALID == 0 {
-            return;
-        }
-        if let Some(count) = unsafe { read_i32(layout.cursor_show_count) } {
-            let flags = FLAG_CURSOR_VALID
-                | if count < 0 { FLAG_CURSOR_HIDDEN } else { 0 };
-            if flags != last.flags {
-                unsafe { publish_cursor(CursorPublished { flags, ..last }, None) };
-            }
-        }
-        return;
-    }
-    unsafe { CURSOR_DIRTY = false };
-    let last = unsafe { CURSOR_PUBLISHED };
-    match unsafe { collect_cursor(layout) } {
-        Ok(state) => {
-            let published = CursorPublished {
-                flags: FLAG_CURSOR_VALID
-                    | if state.hidden { FLAG_CURSOR_HIDDEN } else { 0 },
-                hash: state.hash,
-                hotspot_x: state.hotspot_x,
-                hotspot_y: state.hotspot_y,
-            };
-            if published != last {
-                // Show/hide moves the flags alone, and the region already holds
-                // the bitmap `published.hash` names, so skip the 4 KB rewrite.
-                let bitmap = last.flags & FLAG_CURSOR_VALID == 0
-                    || published.hash != last.hash
-                    || published.hotspot_x != last.hotspot_x
-                    || published.hotspot_y != last.hotspot_y;
-                unsafe { publish_cursor(published, bitmap.then_some(state.source)) };
-            }
-        }
-        Err(flags) => {
-            if flags != last.flags {
-                unsafe { publish_cursor(CursorPublished { flags, ..last }, None) };
-            }
-        }
-    }
-}
-
-// The region comes from the game's allocator, so clear it before the renderer
-// can observe it.
-unsafe fn clear_cursor() {
-    let cursor = unsafe { CURSOR_PTR as *mut CursorSnapshot };
-    unsafe {
-        write_volatile(&mut (*cursor).generation, 0);
-        write_volatile(&mut (*cursor).width, 0);
-        write_volatile(&mut (*cursor).height, 0);
-        write_volatile(&mut (*cursor).hotspot_x, 0);
-        write_volatile(&mut (*cursor).hotspot_y, 0);
-        write_volatile(&mut (*cursor).pixel_hash, 0);
-    }
-    for index in 0..6 {
-        unsafe { write_volatile(&mut (*cursor).reserved[index], 0) };
-    }
-    for index in 0..CURSOR_WORDS {
-        unsafe { write_volatile(&mut (*cursor).pixels[index as usize], 0) };
     }
 }
 
@@ -988,62 +432,32 @@ pub unsafe extern "C" fn companion_init(
         || layout.show_hero_panel_message == 0
         || layout.hide_hero_panel_message == layout.show_hero_panel_message
         || (features & FEATURE_TOOLBOX_FOUNDATION != 0
-            && (layout.prop_context_slot & 3 != 0
-                || !contains(layout.prop_context_slot, 4)))
+            && (layout.prop_context_slot & 3 != 0 || !contains(layout.prop_context_slot, 4)))
     {
         return 0;
     }
     unsafe {
         SNAPSHOT_PTR = snapshot_ptr;
-        CURSOR_PTR = cursor_ptr;
-        TOOLBOX_PTR = toolbox_ptr;
         LAYOUT = layout;
         FEATURES = features;
         INITIALIZED = true;
         TICK_COUNT = 0;
         SEQUENCE = 0;
-        CURSOR_SEQUENCE = 0;
-        CURSOR_GENERATION = 0;
-        CURSOR_PUBLISHED = CursorPublished::EMPTY;
-        CURSOR_DIRTY = true;
-        CURSOR_EVENT_COUNT = 0;
-        TOOLBOX_SEQUENCE = 0;
-        PLAYER_CHAT_COUNT = 0;
-        HERO_COUNT = 0;
-        FIRST_HERO_ID = 0;
-        FIRST_HERO_AGENT_ID = 0;
-        PANEL_STATE = PANEL_UNKNOWN;
-        PANEL_PENDING = PANEL_UNKNOWN;
-        COMMAND_REQUEST = 0;
-        COMMAND_COMPLETE = 0;
-        COMMAND_STATUS = COMMAND_IDLE;
         if features & FEATURE_TARGET_READOUT != 0 {
             publish(State::empty());
         }
         if features & FEATURE_NATIVE_CURSOR != 0 {
-            clear_cursor();
-            publish_cursor(CursorPublished::EMPTY, None);
+            cursor::initialize(cursor_ptr);
         }
         if features & FEATURE_TOOLBOX_FOUNDATION != 0 {
-            let words = TOOLBOX_BYTES / 4;
-            for index in 0..words {
-                write_volatile((toolbox_ptr + index * 4) as *mut u32, 0);
-            }
-            publish_toolbox();
+            toolbox::initialize(toolbox_ptr);
         }
     }
     1
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn companion_dispatch(
-    kind: u32,
-    a: u32,
-    b: u32,
-    c: u32,
-    d: u32,
-    e: u32,
-) {
+pub unsafe extern "C" fn companion_dispatch(kind: u32, a: u32, b: u32, c: u32, d: u32, e: u32) {
     match kind {
         DISPATCH_TICK => {
             unsafe { tick_original(a) };
@@ -1064,41 +478,31 @@ pub unsafe extern "C" fn companion_dispatch(
                 }
             }
             if features & FEATURE_TOOLBOX_FOUNDATION != 0 {
-                unsafe { tick_foundation(layout, state) };
+                unsafe { toolbox::tick(layout, state) };
             }
             if features & FEATURE_NATIVE_CURSOR != 0 {
-                unsafe { tick_cursor(layout) };
+                unsafe { cursor::tick(layout) };
             }
         }
         DISPATCH_CURSOR => {
             unsafe { cursor_original(a, b, c, d, e) };
             if unsafe { INITIALIZED } && unsafe { FEATURES } & FEATURE_NATIVE_CURSOR != 0 {
                 unsafe {
-                    CURSOR_EVENT_COUNT = CURSOR_EVENT_COUNT.saturating_add(1);
-                    CURSOR_DIRTY = true;
+                    cursor::mark_dirty();
                     if FEATURES & FEATURE_TOOLBOX_FOUNDATION != 0 {
-                        publish_toolbox();
+                        toolbox::publish_cursor_event();
                     }
                 }
             }
         }
         DISPATCH_UI => {
             unsafe { ui_original(a, b, c) };
-            if !unsafe { INITIALIZED }
-                || unsafe { FEATURES } & FEATURE_TOOLBOX_FOUNDATION == 0
-            {
+            if !unsafe { INITIALIZED } || unsafe { FEATURES } & FEATURE_TOOLBOX_FOUNDATION == 0 {
                 return;
             }
             unsafe {
                 let layout = LAYOUT;
-                if a == layout.player_chat_message {
-                    PLAYER_CHAT_COUNT = PLAYER_CHAT_COUNT.saturating_add(1);
-                } else if b == FIRST_HERO_ID && a == layout.hide_hero_panel_message {
-                    PANEL_STATE = PANEL_HIDDEN;
-                } else if b == FIRST_HERO_ID && a == layout.show_hero_panel_message {
-                    PANEL_STATE = PANEL_SHOWN;
-                }
-                publish_toolbox();
+                toolbox::observe_ui(layout, a, b);
             }
         }
         _ => {}
@@ -1107,24 +511,10 @@ pub unsafe extern "C" fn companion_dispatch(
 
 #[no_mangle]
 pub unsafe extern "C" fn companion_set_first_hero_panel(shown: u32) -> u32 {
-    if !unsafe { INITIALIZED }
-        || unsafe { FEATURES } & FEATURE_TOOLBOX_FOUNDATION == 0
-        || shown > 1
-        || unsafe { PANEL_PENDING } != PANEL_UNKNOWN
-    {
+    if !unsafe { INITIALIZED } || unsafe { FEATURES } & FEATURE_TOOLBOX_FOUNDATION == 0 {
         return 0;
     }
-    let mut request = unsafe { COMMAND_REQUEST }.wrapping_add(1);
-    if request == 0 {
-        request = 1;
-    }
-    unsafe {
-        COMMAND_REQUEST = request;
-        COMMAND_STATUS = COMMAND_IDLE;
-        PANEL_PENDING = if shown == 1 { PANEL_SHOWN } else { PANEL_HIDDEN };
-        publish_toolbox();
-    }
-    request
+    unsafe { toolbox::request_first_hero_panel(shown) }
 }
 
 #[no_mangle]
@@ -1155,7 +545,7 @@ pub extern "C" fn companion_toolbox_bytes() -> u32 {
 #[no_mangle]
 pub unsafe extern "C" fn companion_cursor_event_count() -> u32 {
     if unsafe { INITIALIZED } {
-        unsafe { CURSOR_EVENT_COUNT }
+        unsafe { cursor::event_count() }
     } else {
         0
     }
