+++ src/renderer/enhancements.ts
@@ -4,168 +4,20 @@ import { createTargetReadout } from "./enhancement-readout.js";
 import { installCursorRefresh } from "./cursor-refresh.js";
 import { createToolboxFoundation } from "./toolbox-foundation.js";
 import {
-  readCompanionSnapshot,
-  readCompanionToolbox,
   COMPANION_CURSOR_BYTES,
   COMPANION_SNAPSHOT_BYTES,
   COMPANION_TOOLBOX_BYTES,
 } from "./companion-snapshot.js";
+import {
+  observeCompanion,
+  recordCompanionLifecycle,
+} from "./companion-observer.js";
+import { decodeEnhancementManifest } from "./enhancement-manifest.js";
 
 const ENHANCEMENT_FEATURE_NATIVE_CURSOR = 1 << 0;
 const ENHANCEMENT_FEATURE_TARGET_READOUT = 1 << 1;
 const ENHANCEMENT_FEATURE_TOOLBOX_FOUNDATION = 1 << 2;
-const ENHANCEMENT_TRANSFORM_ABI = 6;
 const COMPANION_ABI = 3;
-const CONFIG_WORDS = 40;
-
-/**
- * The five values the installer needs out of the kernel's manifest section.
- * The section is JSON the decoder does not control, so it is read as unknown
- * fields and named a manifest only once every one of them has been checked.
- */
-type EnhancementManifest = Readonly<{
-  buildId: number;
-  programId: number;
-  tableSlot: number;
-  configWords: readonly number[];
-}>;
-
-function decodeManifest(module: WebAssembly.Module): EnhancementManifest | null {
-  const sections = WebAssembly.Module.customSections(module, "enhancement_manifest");
-  if (sections.length !== 1) return null;
-  try {
-    const value: Record<string, unknown> | null = JSON.parse(
-      new TextDecoder().decode(sections[0]),
-    );
-    if (value === null) return null;
-    const { buildId, programId, tableSlot, configWords } = value;
-    const hooks = value.hooks as Record<string, unknown> | null;
-    const messages = value.messages as Record<string, unknown> | null;
-    const tick = hooks?.tick as Record<string, unknown> | null;
-    const cursor = hooks?.cursor as Record<string, unknown> | null;
-    const ui = hooks?.ui as Record<string, unknown> | null;
-    if (
-      value.transformAbi !== ENHANCEMENT_TRANSFORM_ABI
-      || !Number.isSafeInteger(buildId)
-      || Number(buildId) <= 0
-      || !Number.isSafeInteger(programId)
-      || Number(programId) <= 0
-      || !Number.isSafeInteger(tableSlot)
-      || Number(tableSlot) < 0
-      || !Array.isArray(configWords)
-      || configWords.length !== CONFIG_WORDS
-      || configWords.some(
-        (word: unknown) =>
-          !Number.isInteger(word)
-          || Number(word) < 0
-          || Number(word) > 0xffff_ffff,
-      )
-      || !tick
-      || !cursor
-      || !ui
-      || JSON.stringify(tick.params) !== JSON.stringify(["i32"])
-      || JSON.stringify(cursor.params)
-        !== JSON.stringify(["i32", "i32", "i32", "i32", "i32"])
-      || JSON.stringify(ui.params) !== JSON.stringify(["i32", "i32", "i32"])
-      || !messages
-      || ![messages.playerChat, messages.hideHeroPanel, messages.showHeroPanel]
-        .every((message) =>
-          Number.isInteger(message)
-          && Number(message) > 0
-          && Number(message) <= 0xffff_ffff)
-      || configWords[37] !== messages.playerChat
-      || configWords[38] !== messages.hideHeroPanel
-      || configWords[39] !== messages.showHeroPanel
-    ) {
-      return null;
-    }
-    return Object.freeze({
-      buildId: Number(buildId),
-      programId: Number(programId),
-      tableSlot: Number(tableSlot),
-      configWords: configWords.map(Number),
-    });
-  } catch {
-    return null;
-  }
-}
-
-function recordLifecycle(state: CompanionState) {
-  if (state.status === "ready") {
-    window.gwAutomation?.set(
-      state.instanceType === 1 ? "game.explorable" : "game.outpost",
-    );
-  } else if (state.reason === "loading") {
-    window.gwAutomation?.set("game.loading");
-  } else if (state.status === "unsupported") {
-    window.gwAutomation?.set("enhancement.unsupported");
-  }
-}
-
-/** The per-frame counters and gauges the observer keeps on the runtime. */
-type SnapshotObserverTarget = {
-  memory: WebAssembly.Memory;
-  snapshotPointer: number;
-  toolboxPointer: number;
-  snapshotReads: number;
-  rejectedSnapshots: number;
-  hertz: number;
-  lastRenderUs: number;
-  renderSamples: number[];
-};
-
-function observeSnapshots(
-  runtime: SnapshotObserverTarget,
-  cursor: ReturnType<typeof createCursorConsumer> | null,
-  readout: ReturnType<typeof createTargetReadout> | null,
-  toolbox: ReturnType<typeof createToolboxFoundation> | null,
-  observeState: boolean,
-) {
-  let frame = 0;
-  let cadenceAt = performance.now();
-  let cadenceTick = 0;
-  const observe = () => {
-    if (observeState) {
-      const started = performance.now();
-      const state = readCompanionSnapshot(
-        runtime.memory.buffer,
-        runtime.snapshotPointer,
-      );
-      recordLifecycle(state);
-      runtime.snapshotReads += 1;
-      if (
-        ("reason" in state && state.reason === "writing")
-        || ("reason" in state && state.reason === "snapshot")
-      ) {
-        runtime.rejectedSnapshots += 1;
-      }
-      window.gwCompanionState = state;
-      const now = performance.now();
-      if (state.status === "ready" && now - cadenceAt >= 1_000) {
-        runtime.hertz =
-          ((state.tickCount - cadenceTick) * 1_000) / (now - cadenceAt);
-        cadenceAt = now;
-        cadenceTick = state.tickCount;
-      }
-      runtime.lastRenderUs = (performance.now() - started) * 1_000;
-      runtime.renderSamples.push(runtime.lastRenderUs);
-      if (runtime.renderSamples.length > 240) runtime.renderSamples.shift();
-      readout?.update(state);
-    }
-    if (toolbox) {
-      const state = readCompanionToolbox(
-        runtime.memory.buffer,
-        runtime.toolboxPointer,
-      );
-      toolbox.update(state);
-    }
-    // Outside the measured window: lastRenderUs stays the snapshot read cost.
-    cursor?.poll();
-    frame = requestAnimationFrame(observe);
-  };
-  frame = requestAnimationFrame(observe);
-  return () => cancelAnimationFrame(frame);
-}
 
 export async function installEnhancements(
   instance: WebAssembly.Instance,
@@ -184,7 +36,7 @@ export async function installEnhancements(
     | (foundation ? ENHANCEMENT_FEATURE_TOOLBOX_FOUNDATION : 0);
   if (featureFlags === 0) return null;
 
-  const manifest = decodeManifest(module);
+  const manifest = decodeEnhancementManifest(module);
   const exports = instance?.exports;
   if (
     !manifest
@@ -193,12 +45,15 @@ export async function installEnhancements(
     || typeof exports?.malloc !== "function"
     || typeof exports?.free !== "function"
     || typeof exports?.enhancement_tick_original !== "function"
+    || exports.enhancement_tick_original.length !== 1
     || typeof exports?.enhancement_cursor_original !== "function"
+    || exports.enhancement_cursor_original.length !== 5
     || typeof exports?.enhancement_ui_original !== "function"
+    || exports.enhancement_ui_original.length !== 3
     || !(exports?.enhancement_hook_slot instanceof WebAssembly.Global)
   ) {
     window.gwCompanionState = Object.freeze({ status: "unsupported" });
-    recordLifecycle(window.gwCompanionState);
+    recordCompanionLifecycle(window.gwCompanionState);
     return null;
   }
 
@@ -208,8 +63,20 @@ export async function installEnhancements(
   // as the bare `Function`, so the kernel's ABI has to be named here or the five
   // call sites below stop checking what they pass.
   const free = exports.free as (pointer: number) => void;
+  if (manifest.tableSlot >= table.length) {
+    throw new Error(`Enhancement table slot ${manifest.tableSlot} is out of bounds`);
+  }
   if (table.get(manifest.tableSlot) !== null) {
     throw new Error(`Enhancement table slot ${manifest.tableSlot} is occupied`);
+  }
+  if (hookSlot.value !== 0) {
+    throw new Error("Enhancement hook is already enabled");
+  }
+  try {
+    // Assignment is the WebAssembly API's only mutability test for a Global.
+    hookSlot.value = 0;
+  } catch {
+    throw new Error("Enhancement hook global is immutable");
   }
 
   let snapshotPointer = 0;
@@ -222,6 +89,32 @@ export async function installEnhancements(
   let disposeToolbox = () => {};
   let disposeCursorRefresh = () => {};
   let installedCallback: CallableFunction | null = null;
+  let installedRuntime: object | null = null;
+  let cleaned = false;
+  const cleanup = () => {
+    if (cleaned) return;
+    cleaned = true;
+    // Disable dispatch before releasing any callback-owned state.
+    hookSlot.value = 0;
+    stopObserver();
+    disposeCursorRefresh();
+    disposeCursor();
+    disposeReadout();
+    disposeToolbox();
+    if (
+      installedCallback !== null
+      && table.get(manifest.tableSlot) === installedCallback
+    ) {
+      table.set(manifest.tableSlot, null);
+    }
+    if (toolboxPointer) free(toolboxPointer);
+    if (cursorPointer) free(cursorPointer);
+    if (configPointer) free(configPointer);
+    if (snapshotPointer) free(snapshotPointer);
+    if (window.gwCompanionRuntime === installedRuntime) {
+      window.gwCompanionRuntime = null;
+    }
+  };
   try {
     if (observeState) {
       snapshotPointer = Number(exports.malloc(COMPANION_SNAPSHOT_BYTES));
@@ -388,9 +281,10 @@ export async function installEnhancements(
           : 0;
       },
     };
+    installedRuntime = runtime;
     window.gwCompanionInstallations = runtime.installation;
     window.gwCompanionRuntime = runtime;
-    stopObserver = observeSnapshots(
+    stopObserver = observeCompanion(
       runtime,
       cursor,
       readout,
@@ -399,43 +293,11 @@ export async function installEnhancements(
     );
     hookSlot.value = manifest.tableSlot + 1;
 
-    const teardown = () => {
-      hookSlot.value = 0;
-      stopObserver();
-      disposeCursorRefresh();
-      disposeCursor();
-      disposeReadout();
-      disposeToolbox();
-      if (table.get(manifest.tableSlot) === installedCallback) {
-        table.set(manifest.tableSlot, null);
-      }
-      if (toolboxPointer) free(toolboxPointer);
-      if (cursorPointer) free(cursorPointer);
-      free(configPointer);
-      if (snapshotPointer) free(snapshotPointer);
-      window.gwCompanionRuntime = null;
-    };
-    window.addEventListener("pagehide", teardown, { once: true });
+    window.addEventListener("pagehide", cleanup, { once: true });
     console.info(`[enhancement] installed for client build ${manifest.buildId}`);
     return runtime;
   } catch (error) {
-    hookSlot.value = 0;
-    stopObserver();
-    disposeCursorRefresh();
-    disposeCursor();
-    disposeReadout();
-    disposeToolbox();
-    if (
-      installedCallback !== null
-      && table.get(manifest.tableSlot) === installedCallback
-    ) {
-      table.set(manifest.tableSlot, null);
-    }
-    if (toolboxPointer) free(toolboxPointer);
-    if (cursorPointer) free(cursorPointer);
-    if (configPointer) free(configPointer);
-    if (snapshotPointer) free(snapshotPointer);
-    window.gwCompanionRuntime = null;
+    cleanup();
     window.gwCompanionState = Object.freeze({
       status: "error",
       reason: error instanceof Error ? error.message : String(error),
