+++ tests/packaged-enhancement-runtime.ts
@@ -574,9 +574,19 @@ async function installTargetReadout(page: Page, moduleBytes: Uint8Array) {
           __indirect_function_table: table,
           malloc,
           free,
-          enhancement_tick_original: () => undefined,
-          enhancement_cursor_original: () => undefined,
-          enhancement_ui_original: () => undefined,
+          enhancement_tick_original: (context: number) => void context,
+          enhancement_cursor_original: (
+            a: number,
+            b: number,
+            c: number,
+            d: number,
+            e: number,
+          ) => void [a, b, c, d, e],
+          enhancement_ui_original: (
+            message: number,
+            wparam: number,
+            lparam: number,
+          ) => void [message, wparam, lparam],
           enhancement_hook_slot: hookSlot,
         },
       },
@@ -788,9 +798,19 @@ async function assertRollbackAfterTablePublication() {
               __indirect_function_table: table,
               malloc,
               free,
-              enhancement_tick_original: () => undefined,
-              enhancement_cursor_original: () => undefined,
-              enhancement_ui_original: () => undefined,
+              enhancement_tick_original: (context: number) => void context,
+              enhancement_cursor_original: (
+                a: number,
+                b: number,
+                c: number,
+                d: number,
+                e: number,
+              ) => void [a, b, c, d, e],
+              enhancement_ui_original: (
+                message: number,
+                wparam: number,
+                lparam: number,
+              ) => void [message, wparam, lparam],
               enhancement_hook_slot: hookSlot,
             },
           },
